+++ src/data.h
@@ -22,7 +22,7 @@ char *data_load_file(char *, char *);
 void data_dir_entries(char *dir, int *, int *);
 void data_archive_entries(char *arc, int *, int *);
 char *data_preproc(char *);
-void data_export_kernel(float *, farray_t *, char *);
+void data_export_kernel(double *, farray_t *, char *);
 void data_export_proto(proto_t *, farray_t *, char *);
